venvs: fix CLI issue bug #2820
https://bugs.linaro.org/show_bug.cgi?id=2820

For a project built around APIs, OpenStack has completely failed to come
up with a rational way of talking to them. This ensures we are actually
building the clients we intended but failed to because of an errant git
checkout command. It also pins down a couple of dependencies we had that
weren't enforced by requirements.txt files.

Change-Id: I30fbb3623f84941d93effe3396e93387fa25a413