ansible nics: use inventory_hostname
In the US cloud in Austin the network controller's primary name is
"r2-a1". However, it has a DNS alias, us-cloud-network, that is used in
my ansible inventory file. We need to use inventory_hostname to do the
right thing in this role. In the simple case ansible_hostname and
inventory_hostname are the same thing so nothing changes.

Change-Id: Ie2fda6cfe2f5ab0cd9491bf3d46e2ea3cb642036