ansible/horizon: enable horizon to force SSL
The current UK devcloud is working because we have an apache redirect
rule for all http://uk1.linaro.cloud to go to https://uk1.linaro.cloud.
I was trying to set this up in austin and I don't have such a rule
configured. The result was that I was given HTTP urls back from horizon
that didn't work.

This change provides sends a header to the HTTP based end of the proxy
to let it know its actually handling HTTPS traffic.

Change-Id: I01c8a28dc4984d90742390f77bc0eada359857aa